+++ src/pysource_codegen/__init__.py
@@ -2,4 +2,4 @@ from ._codegen import generate
 
 __all__ = ("generate",)
 
-__version__ = "0.5.2"
+__version__ = "0.6.0"